by_indexer/requirements.txt
streamlit
pandas
numpy
requests
json
streamlit_autorefresh
plotly.express